+++ crates/rattler/Cargo.toml
@@ -14,18 +14,13 @@ rustls-tls = ['reqwest/rustls-tls']
 anyhow = "1.0.44"
 apple-codesign = "0.22.0"
 async-compression = { version = "0.3.12", features = ["gzip", "tokio", "bzip2", "zstd"] }
-blake2 = "0.10.6"
 bytes = "1.1.0"
-cache_control = "0.2.0"
 chrono = { version = "0.4.23", default-features = false, features = ["std", "serde", "alloc"] }
 digest = "0.10.6"
 dirs = "4.0.0"
-extendhash = "1.0.9"
 futures = "0.3.17"
 fxhash = "0.2.1"
 hex = "0.4.3"
-humansize = "2.1.3"
-indicatif = { version = "0.17.1", features = ["improved_unicode"] }
 itertools = "0.10.3"
 libc = "0.2"
 libz-sys = { version = "1.1.0", default-features = false, features = ["static"] }
@@ -35,6 +30,7 @@ nom = "7.1.0"
 once_cell = "1.8.0"
 pin-project-lite = "0.2.9"
 rattler_conda_types = { version = "0.1.0", path = "../rattler_conda_types" }
+rattler_digest = { version = "0.1.0", path = "../rattler_digest" }
 rattler_package_streaming = { version = "0.1.0", path = "../rattler_package_streaming", features = ["reqwest", "tokio"] }
 regex = "1.5.4"
 reqwest = { version = "0.11.6", default-features = false, features = ["stream", "json", "gzip"] }
@@ -52,19 +48,10 @@ tracing = "0.1.29"
 url = { version = "2.2.2", features = ["serde"] }
 uuid = { version = "1.3.0", features = ["v4", "fast-rng"] }
 
-[target.'cfg(windows)'.dependencies]
-windows-sys = { version = "0.45.0", features = ["Win32_Storage_FileSystem", "Win32_Foundation", "Win32_System_IO"] }
-
 [dev-dependencies]
 assert_matches = "1.5.0"
-axum = "0.6.2"
-hex-literal = "0.3.4"
-insta = { version = "1.16.0", features = ["yaml"] }
-proptest = "1.0.0"
 rand = "0.8.4"
 rstest = "0.16.0"
-tokio-test = "0.4.2"
-tower-http = { version = "0.3.5", features = ["fs", "compression-gzip", "trace"] }
 tracing-test = { version = "0.2.4" }
 
 [build-dependencies]
