+++ crates/rattler_package_streaming/Cargo.toml
@@ -10,7 +10,6 @@ repository = "https://github.com/mamba-org/rattler"
 license = "BSD-3-Clause"
 
 [dependencies]
-async-trait = "0.1.59"
 thiserror = "1.0.37"
 tar = { version = "0.4.38" }
 bzip2 = { version = "0.4" }
