+++ crates/rattler_repodata_gateway/src/fetch/cache/cache_headers.rs
