+++ crates/rattler_repodata_gateway/src/lib.rs
@@ -1,0 +1,11 @@
+#![deny(missing_docs)]
+
+//! `rattler_repodata_gateway` is a create that provides functionality to interact with Conda
+//! repodata. It currently provides functionality to download and cache `repodata.json` files
+//! through the [`fetch::fetch_repo_data`] function.
+//!
+//! In the future this crate will also provide more high-level functionality to query information
+//! about specific packages from different sources.
+
+pub mod fetch;
+mod utils;
