+++ crates/rattler_repodata_gateway/src/utils/encoding.rs
