+++ crates/rattler_repodata_gateway/src/utils/flock.rs
