+++ crates/rattler_repodata_gateway/src/utils/simple_channel_server.rs
